# schedule-service/docs/swagger.yaml
basePath: /raspyx/api
definitions:
...
  models.AddRoomSchedule:
    properties:
      room_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
...
  models.AddSubjectTypeRequest:
    properties:
      type:
        example: Лекция
        type: string
    type: object
...
  models.AddTeacherSchedule:
    properties:
      teacher_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
  models.Day:
    properties:
      "1":
        items:
          $ref: '#/definitions/models.Pair'
        type: array
      "2":
        items:
          $ref: '#/definitions/models.Pair'
        type: array
      "3":
        items:
          $ref: '#/definitions/models.Pair'
        type: array
      "4":
        items:
          $ref: '#/definitions/models.Pair'
        type: array
      "5":
        items:
          $ref: '#/definitions/models.Pair'
        type: array
      "6":
        items:
          $ref: '#/definitions/models.Pair'
        type: array
      "7":
        items:
          $ref: '#/definitions/models.Pair'
        type: array
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.GetTeacherResponse:
    properties:
      full_name:
        example: ФИО
        type: string
      uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
  models.Group:
    properties:
      number:
        example: 221-352
        type: string
      uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
...
  models.Location:
    properties:
      name:
        example: Автозаводская
        type: string
      uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
...
  models.Pair:
    properties:
      end_date:
        example: "2025-06-01"
        type: string
      group:
        $ref: '#/definitions/models.Group'
      link:
        example: https://online.mospolytech.ru/
        type: string
      location:
        $ref: '#/definitions/models.Location'
      rooms:
        items:
          $ref: '#/definitions/models.Room'
        type: array
      start_date:
        example: "2025-02-01"
        type: string
      subject:
        $ref: '#/definitions/models.Subject'
      subject_type:
        $ref: '#/definitions/models.SubjectType'
      teachers:
        items:
          $ref: '#/definitions/models.GetTeacherResponse'
        type: array
    type: object
  models.ResponseAPI:
    properties:
      errors:
        $ref: '#/definitions/models.ErrorResponse'
      message:
        type: string
      requestId:
        type: string
      result: {}
      success:
        type: boolean
    type: object
  models.Room:
    properties:
      number:
        example: ав4810
        type: string
      uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
...
  models.ScheduleResponse:
    properties:
      schedule_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
  models.Subject:
    properties:
      name:
        example: Иностранный язык
        type: string
      uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
...
  models.SubjectType:
    properties:
      type:
        example: Лекция
        type: string
      uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
  models.SubjectTypeResponse:
    properties:
      subject_type_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
...
  models.UpdateScheduleRequest:
    properties:
      end_date:
        example: "2025-12-31"
        type: string
      end_time:
        example: "10:30:00"
        type: string
      group_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
      is_session:
        example: false
        type: boolean
      link:
        example: https://rasp.dmami.ru/
        type: string
      location_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
      rooms_uuid:
        items:
          $ref: '#/definitions/models.AddRoomSchedule'
        type: array
      start_date:
        example: "2025-09-01"
        type: string
      start_time:
        example: "09:00:00"
        type: string
      subject_type_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
      subject_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
      teachers_uuid:
        items:
          $ref: '#/definitions/models.AddTeacherSchedule'
        type: array
      weekday:
        example: 1
        type: integer
    type: object
...
  models.Week:
    properties:
      friday:
        $ref: '#/definitions/models.Day'
      monday:
        $ref: '#/definitions/models.Day'
      saturday:
        $ref: '#/definitions/models.Day'
      thursday:
        $ref: '#/definitions/models.Day'
      tuesday:
        $ref: '#/definitions/models.Day'
      wednesday:
        $ref: '#/definitions/models.Day'
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for schedules
  title: Raspyx
  version: 2.0.0
paths:
...
  /v2/schedule/{schedule_uuid}:
    delete:
      consumes:
      - application/json
      description: Удаление расписания из системы по его идентификатору
      parameters:
      - description: Идентификатор расписания
        in: path
        name: schedule_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Расписание успешно удалено
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.ScheduleResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "404":
          description: Расписание не найдено
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при удалении расписания
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Удаление расписания
      tags:
      - schedule
    put:
      consumes:
      - application/json
      description: Обновление информации о расписании по его идентификатору
      parameters:
      - description: Идентификатор расписания
        in: path
        name: schedule_uuid
        required: true
        type: string
      - description: Данные для обновления расписания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Данные о расписании успешно обновлены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.ScheduleResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "404":
          description: Расписание не найдено
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при обновлении расписания
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Обновление расписания
      tags:
      - schedule
  /v2/schedule/group_number/{group_number}:
    get:
      consumes:
      - application/json
      description: Получение информации о расписании по его group_number
      parameters:
      - description: Номер учебной группы
        in: path
        name: group_number
        required: true
        type: string
      - description: Флаг сессии
        in: query
        name: is_session
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Данные о расписании успешно получены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.Week'
              type: object
        "404":
          description: Расписание не найдено
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      summary: Получение расписания по номеру группы
      tags:
      - schedule
  /v2/schedule/teacher_fio/{teacher_fio}:
    get:
      consumes:
      - application/json
      description: Получение информации о расписании по его teacher_fio
      parameters:
      - description: ФИО преподавателя
        in: path
        name: teacher_fio
        required: true
        type: string
      - description: Флаг сессии
        in: query
        name: is_session
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Данные о расписании успешно получены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.Week'
              type: object
        "404":
          description: Расписание не найдено
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      summary: Получение расписания по ФИО преподавателя
      tags:
      - schedule
  /v2/subject_types:
    get:
      consumes:
      - application/json
      description: Получение списка всех типов дисциплины в системе
      produces:
      - application/json
      responses:
        "200":
          description: Данные о типах дисциплины успешно получены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/models.SubjectType'
                  type: array
              type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Получение типов дисциплины
      tags:
      - subject_types
    post:
      consumes:
      - application/json
      description: Создание нового типа дисциплины
      parameters:
      - description: Данные нового типа дисциплины
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddSubjectTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Тип дисциплины успешно создан
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.SubjectTypeResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при создании типа дисциплины
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Добавление типа дисциплины
      tags:
      - subject_types
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"